[Silver III] Title: 주유소, Time: 356 ms, Memory: 36168 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -9,32 +9,32 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         n = Integer.parseInt(br.readLine());
-        distances = new int[n - 1];
-        prices = new int[n];
+        distances = new long[n - 1];
+        prices = new long[n];
 
         StringTokenizer st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n - 1; i++) {
-            distances[i] = Integer.parseInt(st.nextToken());
+            distances[i] = Long.parseLong(st.nextToken());
         }
 
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n; i++) {
-            prices[i] = Integer.parseInt(st.nextToken());
+            prices[i] = Long.parseLong(st.nextToken());
         }
     }
 
     static int n;
-    static int[] distances, prices;
+    static long[] distances, prices;
 
     public static void main(String[] args) throws Exception {
         input();
 
-        int tmpDistance = 0;
-        int result = 0;
-        int min = prices[0];
+        long tmpDistance = 0;
+        long result = 0;
+        long min = prices[0];
         for (int i = 1; i < n; i++) {
             tmpDistance += distances[i - 1];
-            int price = prices[i];
+            long price = prices[i];
             if (price < min) {
                 result += min * tmpDistance;
                 min = price;
